backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
httpx==0.27.0
anthropic==0.39.0
numpy==1.26.4
pydantic==2.9.0
python-dotenv==1.0.1
